[FIX] account_reports: allow use of deprecated accounts while generating the grouped deferral entries
Currently, if a user has an active deferred entry whose account has been deprecated in the mean time,
we create a move line containing this deprecated account which raises an unclear error to the user.

Now, instead, we let the user create deferral entries for deprecated accounts.

We also add a test to check that old moves are not taken into account when generating deferral entries.

task-id 3544877

X-original-commit: 79d009233b5dc3dbc95008f1828f0653a7e39c58
Part-of: odoo/enterprise#49880