[IMP] l10n_be_hr_payroll: xlsx exp - period change
This commit adds the export to xlsx for the social security certificate
and the social_balance_sheet for the l10n belgian payroll.
I split the pdf export method to not have to copy paste the data
retrieval code and make it easy, if needed in the future, to adapt the
methods to export in the two format at the same time or to change the
data structure without having to retrieve the data 2 times.

In addition to that the reports for the social balance sheet is now
split by a period of one month instead of being done for the whole
period at once. This has be done because the social balance sheet
needs to be done for each month of the period because some data
have to be kept into a certain ratio each month and thus a whole
period ratio is not correct.

task-3506703

Part-of: odoo/enterprise#48787