[FIX] stock_barcode: no backorder for return pickings in barcode
To reproduce:
1. Open picking OUT/0002 in barcode and click return products
2. Set the amount to 10 manually
3. Click validate

Existing behavior:
When trying to confirm a return picking in barcode, depending on the
settings on the specific picking type, the user would sometimes be asked
to create a backorder. For returns, this is not necessary as another
return for the same picking could always be generated instead of
creating the backorder.

Expected behavior:
This commit ignores the setting of the 'create_backorder' option on the
picking type when confirming a return picking in barcode and never
generates or asks for a backorder in this case.

closes odoo/enterprise#49755

Task: 3548962
X-original-commit: 172a212d5a0eed93037ecbd33027a5219eb35316
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Claeys Pieter (clpi) <clpi@odoo.com>