[FIX] web_grid: use context in the action when create record in grid view
Before this commit, when a default value is defined in the context of the
action and the user creates a record in the grid view. The default value set
in the action context is not taken into account.

This commit ensures the action context is also taken into account in the form
view dialog when the user clicks on `Add a line` or `New` buttons inside the
grid view.

task-3504557

closes odoo/enterprise#56530

X-original-commit: 25ecb3ca6497bc14784c54a8df703645f3fa3d5b
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>