[FIX] web_grid: use context in the action when create record in grid view
Before this commit, when a default value is defined in the context of the
action and the user creates a record in the grid view. The default value set
in the action context is not taken into account.

This commit ensures the action context is also taken into account in the form
view dialog when the user clicks on `Add a line` or `New` buttons inside the
grid view.

task-3504557

closes odoo/enterprise#56656

X-original-commit: a558f57a24c058ab8d785d27f353c1d9e40ffc51
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>