[IMP] industry_fsm_report: improve the onboarding for the worksheet
template use case

This commit's purpose is to improve the onboarding for the worksheet
template use case. The idea is to give the user the opprtunity to create
a new template if none are available on the db.
The user needs to have admin rights, otherwise access right error will
be triggered on the template creation. There also should be no worksheet
in the db, and only one 'default worksheet' template. If there are other
data, we can assume that the user know how to use the feature and
doesn't need help to discover it.

details of the implementation:
A wizard was added when the worksheet button is used on the task form
view. This wizard will only be displayed when the above conditions are
met. Some test were added.

task-3377213
https://www.odoo.com/web#id=3377213&menu_id=4720&cids=1&action=333&active_id=4105&model=project.task&view_type=form

closes odoo/enterprise#48914

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>