[IMP] appointment: allow multiple users on custom appointments
Remove the constraint preventing to set more than one user
on a 'custom' appointment_type. While creating it
from the calendar is linked to the logged user, one could also
modify it or create it from the invitation submenu, and set
multiple users. The arbitrary constraint is removed.

Tests are adapted accordingly.

Task-3549573

X-original-commit: d943ffc5d621b279fab9c98fa8b4d11e32a2e570
Part-of: odoo/enterprise#53288