[IMP] appointment: allow multiple users on custom appointments
Remove the constraint preventing to set more than one user
on a 'custom' appointment_type. While creating it
from the calendar is linked to the logged user, one could also
modify it or create it from the invitation submenu, and set
multiple users. The arbitrary constraint is removed.

Tests are adapted accordingly.

Task-3549573

Part-of: odoo/enterprise#49867