[FIX] *: Accounting tab always shown on product form view
*: account_accountant

Description of the issue/feature this commit addresses:

Wrong visibility assignment of the "Accounting" page of a product's
form view.

Desired behavior after this commit is merged:

This commit adds a group necessary for the view of specific content inside
the said page after it was removed from the default groups of the page in
the account module's version of the page.

Adding this commit, the right groups are set on the Accounting page of the
product form view when account_accountant is installed.

task-3525878

closes odoo/enterprise#48843

Related: odoo/odoo#138483
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>