[FIX] social_demo: remove redundant state update in linkedin
Before this commit:
When a user tries to post a new post with demo data of LinkedIn, the chatter is
also logged twice.

Reason:
The state is updated twice (in the _post_linkedin and _post_demo methods) and
due to that, the _check_post_completion method is called twice.
We have solved this issue by removing the state update in the _post_linkedin
method.

After this commit:
The Chatter will be only logged once.

Task-3459739

closes odoo/enterprise#50217

X-original-commit: 20d86927efc584add36f7b5ae928c9f06fae3bb6
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>