[IMP] documents_account: show embedded PDF inside XML
Purpose
=======
In accounting, some XML invoice documents have PDF embedded in base 64.
We want to show a preview of the embedded PDF, and also to view that
PDF when we click on the preview.

The PDF embedded inside the XML is encoded in base 64.

We support the UBL format
> https://docs.peppol.eu/poacc/billing/3.0/syntax/ubl-invoice

Task-3537562

Part-of: odoo/enterprise#49111