[IMP] industry_fsm_stock: display right warehouse in forecast reports
In a FSM project, when using the `products` smart button, then
`view availability` in the 3-dots menu, the warehouse displayed was
always the first created one.

The warehouse displayed should be related to which is defined on the
SO or the current user one if missing. If user's default is also
missing, we use the warehouse that have the lowest sequence (and that
belong the the task's company).

To reproduce behavior:
- Install `industry_fsm_stock`
- Create second warehouse (WH2)
- Assign the new warehouse (WH2) to current user
- Open a project in FSM app
- On `Products` smart button, check the availability of a product. The
  displayed forecast report should be the base warehouse (YourCompany)

task-3522292

closes odoo/enterprise#48913

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>