[IMP] industry_fsm_*: add pick up button on fsm task
Currently, the user has to open the SO form view and check the deliveries to identify if they need to bring something and what. This is not the most practical.
In addition, it is common for companies not to give sales access to field workers as they don't want them to see the prices and the margin they are making.

In this PR we'll add a 'x products to pick up' stat button
should represent the following stock move lines
- associated with the SO linked to the task
- and that are not done or canceled
- and the destination location should be a 'customer location' type

Added reading rights on stock.move/stock.picking for Field Service/User group

Task-3511049

closes odoo/enterprise#49074

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>