[IMP] project_enterprise: order smart schedule by deadline
When smart scheduling, prioritize tasks with a deadline over no deadline,
then tasks with a shorter deadline over longer ones.

Don't group by project (not possible anymore + not very useful) to get
the tz, calendar and company. Rather, take the one of the user, or the
project you're in the context of, or the env.

task-3349377

closes odoo/enterprise#48944

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>