[FIX] knowledge: fix some UI and portal changes
Current behavior before this commit:
When a user creates an article in a private section and then publishes it on
the web, now if a portal user accesses that link, the article will appear in
the private section of the portal as well.

Reason:
There was no specific domain to prevent the portal user from showing private
articles of others.

Desired behavior after this commit:
Published web articles will not shown in the portal's private section.

Task-3486083

closes odoo/enterprise#58884

X-original-commit: efba53f6649234c25d1d8646e65925f807cc119c
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>