[IMP] website_sale_renting: use timezone of website in rental order
Before this commit, when user make rental order from website then rental periods
are set in user timezone even if store is located in different timezone.

In this commit, add a new field in website for timezone and use that timezone
when user makes a rental order and displayed date picker according to websites
timezone.

task-3293817

closes odoo/enterprise#52490

Related: odoo/odoo#145705
Signed-off-by: Valentin Chevalier <vcr@odoo.com>