[FIX] web_studio: boolean field draggable when editing form
Prior to this fix, users were unable to drag and drop boolean fields
(e.g., checkboxes) within the form editor in Studio. This issue is
reproducible through the following steps:

- Install Studio
- Go to Contacts (or any other app)
- Select any contact and edit the contact form with studio.
- Now search or add a `boolean` field, usually a checkbox will work.
- Now try to move it with drag and drop from one position in the form to
any other position.

This problem was introduced in the commit `d8958a4`, where a condition
was added to the rendering logic of the fields. This condition
inadvertently skipped boolean fields, preventing them from receiving
the necessary attributes for drag-and-drop functionality.

The fix adds an XPath expression specifically targeting boolean
fields that adds the missing attributes to the boolean
fields, like `labelRef`, in order to later get the draggable attribute.

opw-3495765

closes odoo/enterprise#53105

X-original-commit: e9171d52438a15e53d21be358d0e825e98ebcab9
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>