[REF] mrp_workorder: Allow customization of no_start_next value
The no_start_next context key defines if mrp.workorder.do_finish
will return an action to start the next workorder. It is used
when we click the "Mark as Done and Close MO" button to redirect
the user from the tablet view to the workorders kanban view.

If the user wants to process next workorders without having to
get back to the kanban view, he can use "Mark as Done", but
the manufacturing order will not be closed automatically and
he will have to close the MOs manually in the backend.

Having this value hardcoded in the backend did not allow to
customize this behaviour by changing the value of the context
key on the button through an extension view.

closes odoo/enterprise#51791

X-original-commit: 53bf849e12110d8ae18ecb0943e61c2e541e1fda
Signed-off-by: William Henrotin (whe) <whe@odoo.com>