[IMP] web_studio, *: add ThumbnailItem component
*: website_studio

This commit aims to reduce duplication of code in studio, and creates the
ThumbnailItem component. Before this commit, the same exact layout was used
in different places of the UI to show a card with an icon, a name and an
optional 'More' menu at the lower part of the card.

Instead of reusing the same template, and to assure an identical behavior and
style to this element, we added this component instead. The style related to the
feature has been moved to be located next to the component, in its folder. This
folder is located in a 'components' folder under the studio module. Maintaining
the component should be easier.

This commit also reduces the usage of scss mixins, and uses bootstrap classes
instead for many properties that were present in the original files. Tests have
been adapted to this naming change.

task-3572251

Part-of: odoo/enterprise#49809