[FIX] web_studio: fix tabs and columns size issue in field service module.
Steps:

Open the Field Service module.
Go to configuration.
Open worksheet templates.
Select any worksheet template and click on "Design Template".
Drag and drop tabs or columns inside the comments field.
The size of tabs/columns appears very small.

Issue:

When dragging and dropping tabs or columns inside the comments field, the size
of the tabs/columns appears very small.

Cause:

This issue occurs because tabs and columns are allowed inside the comments field.
Inside the comments field, there is a grid, which reduces the size of tabs and
columns.

Fix:

It is unnecessary to allow tabs and columns inside the comments field. We should
restrict them from being placed inside the comments field. Tabs and columns
should be allowed inside sheet and after group.

task-3631285

closes odoo/enterprise#54361

X-original-commit: df60e871178dcc6563240f540be5cdcf8a124199
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>