[REF] spreadsheet_edition: remove SELECT_ODOO_LIST command
There was `SELECT_ODOO_LIST` command that had the list UI plugin store
a list id, which could then be fetched with the getter
`getSelectedListId`.

The getter was only used once (outside the tests): to display the
correct list in the list side panel. But this is not even useful, as
we can just give the correct list as a props of the side panel.

This commit removes the `SELECT_ODOO_LIST` command and its corresponding
getter. It also refactor the list side panel, now there is two
distinct side panels (one for the list of lists, one for the details).
This avoids having to manage a state to determine which list should
be displayed.

Also take the occasion to remove some left-over JQuery and legacy test
utils in the tests.

Task: 3557655
Part-of: odoo/enterprise#49058