[FIX] knowledge: display burger menu on medium screen size
At medium screen sizes, the Knowledge editor and sidebar cannot be
accommodated due to limited space. When attempting to adjust the
sidebar within these screen resolutions, the buttons in the topbar
may extend beyond the boundaries of their parent container. To
enhance the layout, we will introduce the burger menu for medium
screen sizes.

task-3507172

closes odoo/enterprise#51190

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>