[REF] spreadsheet: colocate special args in one place
This is the enterprise counter part of a commmunity commit

For the ODOO.PIVOT.HEADER functions, "special" values such as measure and
total were managed in different places (total management  was also duplicated).

Now the measure and total are managed in one place, in the high level method.
There's also now a dedicated method to get a measure display name. The method
`getGroupByDisplayLabel` was perverted at that purpose.

Task: 3570281
X-original-commit: 7283e55787b282eccbe047fbb90aad53f842d8ba
Part-of: odoo/enterprise#55090