[IMP] planning: make "Shifts in Conflict" filter again
In a previous commit, we customized the filter "Shifts in Conflict"
in Gantt and Calendar views to be able to still see other shifts.
Makes sense: as the idea is to plan them conveniently, you want to
see the rest of the schedule.
So, rather than having a domain filtering out the non-conflicting
shifts, we just display them with a lower opacity.

The issue is that, in other views, the filter has no effect.
With this commit, we filter out the non-conflicting shifts
in List and Kanban views.

task-3555519

Part-of: odoo/enterprise#49104