[IMP] industry_fsm_sale: default invoicing option
When creating a invoice from a field service task, 'regular invoice'
option is set by default. And with some conditions, 'regular invoice'
triggers an error when we try to create a invoice with this option.
Instead of setting by default the 'regular invoice' when we know that
it will trigger an error, we check the condition before displaying the
form view. If it trigger the error, we set by default the 'down payment'
option.

task-3514449

closes odoo/enterprise#49180

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>