[FIX] web_grid: selection field groupBy traceback fix
Steps:
-Install project_timesheet_forecast_sale.
-Open timesheets module.
-GroupBy Billing Type.

Issue:
-If project_timesheet_forecast_sale module is not installed and we groupBy
Billing Type then it works but when this module is installed and then we try
to groupBy Billing Type, a traceback appears.

Cause:
-Billing Type (timesheet_invoice_type) is a selection field and rest all
groupBy are mostly many2one fields.So, when in selection field groupby
value is not passed then None value is passed.

Fix:
-We are checking for valid condition,if there is no value then we are just
passing None.

task-3522253

closes odoo/enterprise#52560

X-original-commit: 94eae276ef64e7f6a69172fd1b66d121bc804155
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>