[IMP] project : default allocated hours
When a user creates a new task from the gantt view, we add default
allocated hours based on the planned dates set by the user and on
the workable hours of the assignee(s) during this interval of time.

It is only a default value set when creating a record, It can be changed
afterwards by the user at anytime.

taskid:3432125

closes odoo/enterprise#49168

Related: odoo/odoo#148875
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>