[IMP] account_reports: Split account names during xlsx print
There are a few reports/options that have the account as line name.
When it happens, the account display name is shown, which contains
both the code and the name.

In this change, we aim to improve the xlsx print by dynamically
splitting the name in such a case in two columns, easing manipulation
of the file after export.

Task id # 3525247

closes odoo/enterprise#50946

X-original-commit: 3d4a289fb7090e0fbd6be11eea01bdab0217de76
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>