[FIX] documents: validate document existence using the `datapoint_id`
Issue:
- The existing condition inside the `insert()` method incorrectly identifies
the incoming new record (data) as an already existing record inside
`this.store.Document.records`

Reason:
- It uses `id` to check for a new document, but the `id` remains the same in
the folowing cases.
   - In case of replacing a document with a new document.
   - Replacing the requested document with the new document.

- As a result the records inside the `this.store.Document` remain inconsistent
with the records actually present in the database.

After this commit:
- We use the `id` of the record (ex- `datapoint_32`) to check for the
presence of a new document.

Refer the related community PR for the issue being tested by the test case in
this commit.

Task-3489455

closes odoo/enterprise#54752

X-original-commit: 0b92ab69fd28821f823bd7da77ce6bb663cd7fdd
Related: odoo/odoo#150257
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>