[FIX] account_asset: reevaluation with a decrease
After a reevaluation with a decrease, we postponed next entries.
It was deemed a faulty behaviour.
We decided to remake the behaviour from 15 that is,
spread the remaining amount on the remaining period.

While fixing this, we decided to fix the issue with lock dates,
and by remaining amounts that were left after a reevaluation.
To do so, we removed the dependency on asset_number_days and on dates.
The problem with asset_number_days was that it is stored as an Integer, while doing computations as if it is a Float (so rounding issues).
With date, it didn't take into account the fact that lock dates postpone the posting.

closes odoo/enterprise#52218

X-original-commit: a53d4b7b4b7715e745dcadeb4c0222c9e7061022
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>