[IMP] spreadsheet_edition: delete history after 1 year
Currently, as time goes, the revision table only grows and grows. Revisions were
previously garbage collected after 3 months, but since the history feature
(task 3142006) they are no longer deleted to allow going back in time.

With this commit, to limit the size of databases, we delete the history
of spreadsheets that have not been modified for more than a year
(overridable with an 'ir.config_parameter')

closes odoo/enterprise#49793

Task: 3562324
X-original-commit: cf98e0dab2999d3f05290d9e0d5c8a96876ac9e6
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>