[FIX] iot: Worldline C++ prototypes fix
This poke fix aims to remove an issue present in current Worldline
interface and driver related to the way we work with strings between
Python and C++.

Currently, we are creating string buffers in Python (char *) and passing
pointers to them (char **) to our C++ Worldline EasyCTEP library.
Then inside our C++ code we copy a string inside that pointer (char **).
This is a mistake which can easily lead to a buffer overflow, we need to copy strings inside the buffers and not pointers to them.

task-3562527

closes odoo/enterprise#49699

X-original-commit: 954fe0af816f68461c345125b51f23b4d1335fb0
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>