[FIX] account_accountant: Notes in bank line not editable
Problem
---------
The "Notes" field in Bank Reconciliation list view (optional_hide)
should be editable. It's not working anymore in 17, it's working in
16.4.

Steps
---------
1. Install `account_accountant`
2. Accounting Dashboard > Bank journal > List view
3. Unhide Note column > Click on any empty notes

Solution
---------
The field was set as invisible when it was empty. However, this appeared
to also hide the editable box from the HTML field. The box was then not
clickable and thus not editable. Removing the invisible attribute solves
the issue while maintaining the idea that nothing should be shown (no
empty box) if there is no note content.
From this fix, another issue appeared back in the Kanban view: the notes
field was editable with a sequencing arrow appearing on the left. We
don't want the note field to be editable in the Kanban view. To solve
the issue, we replace the HtmlField tag in the XML template by a t-out
so that the attribute set on the python fields are used in the view.

task-3537776

closes odoo/enterprise#50375

X-original-commit: fb553764fa14f87bfa778888ec25a054f6f98c87
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>