[FIX] web_enterprise, *: convert `text-odoo` to `text-primary`
This commit aims to fix an issue related to the `text-odoo` class.
That class was generated based on the `$o-theme-text-colors` map, which
included a `"odoo"` value. Since this value was removed, every occurrence
of the class would display a default color for the text, based on the
body font color.

Since the `text-odoo` class was previously using the primary color, we
convert every `text-odoo` occurrence to `text-primary`, to achieve the
expected behavior.

task-3201038

Part-of: odoo/enterprise#49613
Co-authored-by: chgo-odoo <chgo@odoo.com>