[FIX] account_online_synchronization: Have a real asynchronous flow
The aim of this commit is fixing the asynchronous flow to get what we
want since the beginning.
A first version(ed7d7486f30e076e6040a2d528a2c4926225bfa9) of the flow
only adds the call to the cron to make the transactions fetch
asynchronous. This first version doesn't update the status on
the user dashboard. It means that the spin turns until the user
refresh his/her dashboard.
Now the dashboard is automatically updated by websocket once we
receive a response from Odoo Fin (if we receive an error or if we
fetch or not some transactions). We're using the field
connection_state_details on account.online.link (added by this commit:
4e88e4ff87338b36baf317194c1a52ba57da35cd) to save the last connection
state (how many transactions we fetched or the error).
The field is on the wrong model. We probably have to improve that
in master. This commit also adds a new cron for connection with a
fetching_status set to "waiting". This cron should only be used
to fetch transactions asynchronously. When a user receives a new
connection state, it's saved on the connection, once the user opens
the bank statement lines view, this state is "consumed", meaning that
we remove this state from the connection.
This commit also changes how the function "_fetch_transactions"
works. Now, we're trying to refresh all accounts, if we have at least
one account ready to synchronize, we call the cron to fetch transactions
asynchronously.

task-3568712

closes odoo/enterprise#49953

X-original-commit: b8eef19b0d9807fdcd49c48a69a29754a56dfd30
Signed-off-by: William André (wan) <wan@odoo.com>