[FIX] account_reports: properly manage branches in multi-company filter
Example of the issue:

1) Create a company called "main"
2) Create a branch for "main" called "branch"
3) Make an invoice in "main"
4) Make an invoice in "branch"
5) Modify the Aged Receivable report, and set filter_multi_company to 'disabled'
6) Select "main" as current company. By default, "branch" is selected as well, keep it like this.
7) Open the Aged Receivable

====> You only see the invoice made for "main", not the one in "branch". While totally correct in a regular multi-company setup, this is not the behavior we want when those companies contain branches.

When filter_multi_company is set to 'disabled', we also want to include all the branches of the active company that are currently selected in the company selector.

When it is set to 'tax_units' a no tax unit is currently selected, we need a slightly different behavior: indeed, some sub-branches of the active company could submit their tax report separately from their parent, so we don't want to include all of them. To guess what sub-branches fall into that case, we use the fact their VAT number is set to a different value than the active company's. All sub-branches with the same VAT number as the current company that are enabled in the company selector will be included automatically when no tax unit is selected.

X-original-commit: dd3133c5407e8e72557686011bc7615a8d8d2885
Part-of: odoo/enterprise#50243