[FIX] timesheet_grid: fix timesheet timer button stop traceback
Issue:
Encountering a traceback error when attempting to stop the timer on an existing
record in the timesheet kanban view.

Steps to Reproduce:
Navigate to the timesheet kanban view.
Use the group-by search option.
Click on the start timer button for an existing record and stop the timer.

Cause:
The traceback is triggered due to an undefined timesheet value. The cause lies
in a conditional check that restricts setting the timesheet value only for
dynamic records.

Solution:
Remove the conditional check that limits timesheet value assignment to dynamic
records. and add a new condition to ensure the timesheet value is set for both
dynamic and existing records.

task-3555184

closes odoo/enterprise#54010

X-original-commit: c1451916b6a3efa6d0457ea9b91d17705d596159
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>