[FIX] approvals: proper view for approvals product Lines
Steps to reproduce:

- Install approvals.
- Go to the app and create a new approval like `Borrow Items`.
- Now fill the form with atleast 1 product.
- Go to mobile view of the same page and reload.
- See the product Lines

The issue is that since we don't have any kanban view specific for
approvals, we will try to load a generic one which in this case will be
only showing the id of the product. Creating a custom kanban view for
this lines we can do a proper display like displaying the name of the
product and the quantity.

opw-3449483

closes odoo/enterprise#51028

X-original-commit: 7f90aebb4d7e7fcddd46eaf18b7a3d380f7a3e1a
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>