[IMP] account_reports: Show disabled buttons instead of hiding them entirely when not branch_allowed
Before this commit, when a report was opened with branches enabled that did not correspond to the full company branch hierarchy, export buttons were hidden, to avoid the user generating an official export file for a sub-branch instead of the whole group by mistake.

Instead of hiding those buttons, we now display them as disabled in this case, to avoid confusion.

X-original-commit: 7cd7979157dff29ad0f65f81ea30b082657dea1d
Part-of: odoo/enterprise#50243