[IMP] account_report: landscape on multiple report
With the introduction of the 'section' implementation, it's now possible to
print multiple reports in a single batch. However, a problem arises when one of
the reports requires a landscape format (e.g., GL) while the others should
remain in portrait orientation.

This pull request addresses this issue by ensuring that the landscape format is
applied only to the report that specifically requires it. For example, the
Balance Sheet (BS), Profit and Loss (P&L), and Trial Balance (TB) reports will
be printed in portrait, but the General Ledger (GL) report will be printed in
landscape.

Here's what has been done:
For each report, we execute wkhtmltopdf and create a stream for that bytes pdf.
We create a list of these streams and utilize the 'merge_pdfs' function to
create a single stream containing all the PDFs. By passing this combined stream
in the 'file_content' parameter, we achieve the desired outcome of printing a
single PDF file with the correct format for each report.

closes odoo/enterprise#52243

Task: 3566920
X-original-commit: ec33abfa731d593c9b2044a8d15d5de840a3d757
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>