[FIX] delivery_easypost: add `contents_explanation`
Steps to reproduce:
1. Set up UPS with EasyPost shipping integration
2. Make a shipping with destination in Mexico
3. Attempt to validate the delivery

Get an error message:
`"EasyPost returned an error : A package in a Mexico shipment
must have a Merchandise description"`

Based on new Carta Porte regulation in Mexico, (See https://www.easypost.com/blog/2023-02-13-all-you-need-to-know-about-the-new-ups-carta-porte-regulation)
EasyPost enforces some carriers to have `contents_explanaton` field.
This field is a have a package-level description of the shipment.

This commit uses a truncated list of product names to this field so
that the labels can be generated.

opw-3555670

closes odoo/enterprise#51220

X-original-commit: 1bc436213db625ebcba7a7cc72fed067430a6fc1
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Ahmad Khanalizadeh (khah) <khah@odoo.com>