[IMP] appointment: adapt slots changing resources or users
Before, there was a constraint raising when changing
staff_users on an appointment_type if the ones restricted
to slots were not in staff_users anymore.

Instead of raising, and in order to have a similar behavior
for resources, when changing a staff_user or a resource,
synchronize the slots to match that user / resource domain.
The fields of restricted users / resources fields are made
into computed stored fields depending on the appointment
values.

A test is added to ensure this behavior.

Task-3549573

closes odoo/enterprise#53189

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>