[IMP] web_studio: Add dynamic table on report editor
After this commit, there is a new command on the report editor (Dynamic table).
A popover with only the x2many fields is shown and after the user selects the field,
a table with a `t-foreach` is created. Inside of the t-foreach's row one can access the
fields of the co-model.

task-id-3531496

closes odoo/enterprise#50068

X-original-commit: 6fac3cac48b47744167aa3f3ecc5336e920ad3c7
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>