[FIX] knowledge: add missing template filter in article selector
When merging the models `knowledge.article` and `knowledge.article.template`,
we unintentionally forgot to add a condition in the search domain of the
article selection dialog to filter out the templates from the search
results. As a result, the dialog currently lists both articles and
templates. This commit will simply add the missing condition in the
search domain to exclude the templates from the search results.

Steps to reproduce the issue:
1. Open any view
2. Click on the cog icon in the control panel
3. Click on the "Knowledge" section of the dropdown
4. Click on the button "Insert view in article"
=> The modal allowing the user to select an article displays the articles
   and the templates.

TO BE: The modal should only display the articles and not the templates.

Reference: https://github.com/odoo/enterprise/pull/48691

task-3571715

closes odoo/enterprise#49747

X-original-commit: e11ea44edb6e49769335a177a1e41c4d51deca24
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>