[FIX] l10n_es_reports: mod 349: get invoice totals without ignoring negative lines
To reproduce the issue:

1) On a Spanish company, create a customer invoice, so that:
    - it has a line at 100, and another negative one at -42.
    - it uses the default Spanish sales tax of 21% on both lines
    - in is made on a non-Spanish European partner
    - it defines a value in mod 349 type; E, for example
    => the total of the invoice is hence 70.18€

2) Post the invoice

3) Open mod 349 for the period corresponding to the invoice date.

    ===> The invoice is reported with an amount of 112.18€, which is wrong.

This happens because the domains of the report line include lines that have debit>0. In our example, two lines are debit:
    - 70.18, on the receivable account
    - 42, corresponding to the negative line

For these lines, we want to use the total of the invoices/refunds, so instead of just counting on selecting one side of the move (and ignoring the case of the negative lines), we should consider the payable/receivable lines only.

OPW 3559094

closes odoo/enterprise#50051

X-original-commit: f89958b4b549131b7b00dc077136099583a93340
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>