[FIX] web_gantt: excessive rendering of the progress bars
This commit fixes an issue with the gantt progress bars which were
re-rendering every time the user would hover a different cell in the
gantt view even if the cell is on the same row as the one before. The
progress bar component will now only be re-rendered when different rows
are being hovered.

closes odoo/enterprise#49921

Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>