[FIX] core,*: Fix display_name computations for new records.
As of https://github.com/odoo/odoo/pull/114024, `display_name` is
implicit in every form view (see the use of addFieldDependencies in
the Form controller). Therefore, when you create a new record for this
model, it calls the first `onchange`, which will compute display_name
for a new record (id without origin). Some `_compute_display_name`
don't handle new records correctly and raise a traceback. One is
directly accessible:

- Products > Open TaxCloud Category field > New => Traceback

Other models are inaccessible by default (no view to access or create a
new record), but if someone creates a view for them with studio (or
modifies an existing one to allow creation):

- `product.avatax.category`
- `social.post.template`

Change the code of `_compute_display_name` on these models to be more defensive and avoid (potential) tracebacks.

closes odoo/enterprise#50263

X-original-commit: 01454ae62a1fd3c7bec86b7c077d5853a2e73589
Related: odoo/odoo#141181
Signed-off-by: Rémy Voet (ryv) <ryv@odoo.com>