[FIX] web_studio: reportEditorWysiwyg: listen on historyStep (not mutations)
Before this commit, we waited for mutations to take place to determine the state undo/redo
of the reportEditorWysiwyg.
In some cases this was problematic, namely when inserting a field, because the OdooEditor
handled its history on its own during the operation.

After this commit, the reportEditorWysiwyg listens on the "historyStep" event of the OdooEditor, which seems
to be the right API for this.

X-original-commit: d1e965241039c6afa7961cd6f937d1c8b9f64aec
Part-of: odoo/enterprise#50068