[FIX] knowledge: removes comment classes for print
This commit fixes a bug where when a user would try to export an article
the anchors for comments were still stylized.
This led to printing documents with unnecessary noise in the article.

Now, we added the media css query `@media screen` for comment anchors so
that we apply the style when using a screen. Removing the style from the
printed document.

task-3572009

closes odoo/enterprise#50297

X-original-commit: 120fcd57d377af9674806356574544702fcd7c09
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>