[FIX] appointment: set author and email_from as organizer
Before this commit
==================
For public users, we assign 'author_id' as odoobot in appointment booking and
cancellation emails, which can lead to inconsistency since 'author_id' belongs
to odoobot while 'email_from' belongs to the event organizer.

After this commit
==================
Now in all cases, 'author_id' and 'email_from' are set if the organizer exists.

Task-3433753

closes odoo/enterprise#53055

X-original-commit: adfe96612cfd0a0b29c8af7d03b19e87ed14155d
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>