[FIX] appointment: set author and email_from as organizer
Before this commit
==================
For public users, we assign 'author_id' as odoobot in appointment booking and
cancellation emails, which can lead to inconsistency since 'author_id' belongs
to odoobot while 'email_from' belongs to the event organizer.

After this commit
==================
Now in all cases, 'author_id' and 'email_from' are set if the organizer exists.

Task-3433753

closes odoo/enterprise#53017

X-original-commit: ffe75b94e1afe93ce95223c78b8ab82c16973375
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>