[FIX] account_reports: Aged Payable audit cleanup
Description of the issue/feature this commit addresses:

The cleanup of the audit in the Aged Payable report comes from two reasons.

The first is that making the Aged Payable report auditable makes the Aged
Receivable show no cell which is obviously not wanted

The second is that a system for making cells auditable without having to create
a dedicated component has been added in the code and its use makes the audit of
a report much more understandable.

Desired behavior after the commit is merged :

The behavior of the audit on the Aged Payable should not change. However,
the Aged Receivable is not bugged anymore.

X-original-commit: aa580b499c3d8fdb722e85c96f6b989a5bdad521
Part-of: odoo/enterprise#50221