[FIX] hr_appraisal: create appraisals correctly from cron
Purpose:
The appraisal created from cron is already is confirmed state and
it should have same bihavior as manually created and confirmed one.
The feedback templates should be loaded and feedback not published.

Prior this commit:
When we create appraisal from cron, the template is not loaded.
Also both feedbacks are published.

After this commit:
The appraisal that is created from cron, has loaded templates and
employee and manager feedbacks are not published.

task-3573802

closes odoo/enterprise#49861

X-original-commit: 9eb3e7ad5a8d669a532529bf20be9303c7e202f8
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>