[FIX] account_reports: totals below
To add a total below line, we go through all the lines and check if the
current line id does **not** start with the total below parent id line.
If it's the case, we add it before adding the line. This makes sense
because the total below should be the last child of a line.

The problem is that you may have cases where the line id start with the
total below parent id even tho they are not related. For example, if the
line id is something like x|y-14 and the total below parent id is x|y-1.

To fix it, we check if the line id does **not** start with the total
below parent id **ending with** the line id hierarchy separator.

X-original-commit: df39f986d1f6786176291b8f58f620ebb1e911d0
Part-of: odoo/enterprise#50077