[FIX] {account_invoice/hr_expense}_extract: Fix auto-extraction of attachments.
Problem
---------
In the commit 24345812cb015e86f362f8c35176646239bfb803 (odoo), we
modified the extraction of attachment in such a way that mails coming in
a journal from aliases with several attachments create several invoices.
Due to the urgency of the task, the computation of whether an attachment
needs auto-extraction or not, has been skipped. This commit fixes this.

The problem of the failing test came from the conditional order that
verified if the auto-extraction was needed or not. Here is the
difference

What the test expected              | What we had
------------------------------------------------------------------------
1. From alias                       | 1. New move
    a. Attachment is PDF            | 2. From alias
2. New move or purchase document    |     a. Attachment is PDF
                                    | 3. Purchase document

So, new documents were always extracted even though the option for
extracting PDF only was active. Since all document created through a
mail from a mail alias are new document, all attachments (no matter if
binary or PDF) were extracted.

Solution
---------
Update the `_needs_auto_extract` and decoder to behave as the test
expects; by shifting the conditions around. Also context key checks in
both the `_needs_auto_extract` method and decoder to completely disable
the calls to the OCR when needed.

closes odoo/enterprise#54482

X-original-commit: 1a4808a63994006920e3150d84ba775a7742f70c
Related: odoo/odoo#149693
Signed-off-by: Laurent Smet (las) <las@odoo.com>