[IMP] account_reports: display companies in the UI
When having multi company (with branches and tax units) it may sometimes be a
bit difficult for the user to realize what companies are currently used by the
report.

This commit adds under the filters, on top of the report the names of the
companies present in the options dict with the key companies under certain
conditions.

closes odoo/enterprise#50219

Task: 3573675
X-original-commit: ce03b1efe52c65dd3821effe98f620702c0667e7
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>