[FIX] stock_barcode: hide return button when appropriate
Fixes 2 cases of the return button in barcode showing when it shouldn't.
- when in the "info" view of the picking then the button was showing on
  top of the chatter when it shouldn't show in this view at all
- in done manufacturing orders

Note that a getter isn't added to the JS to handle whether or not the
button should be shown because mrp_model extends the picking_model and
batch pickings do not currently offer the "Return" option in the
backend, so it didn't seem necessary to create a getter to only override
it to be "False" for mrp_model only.

closes odoo/enterprise#49899

X-original-commit: 1b524eb68c6ee4b80bc3ec2cf561a8966d000ed0
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>