[FIX] planning, sign: list renderer alignment top
Prior to this commit some td in the list renderer were vertically align
on the middle while other are aligned on top. This comes from commit[1]

This commit adapts the wrench icon in planning to be centered vertically
with the new  `$o-line-size` variable introduced in community.

The new variable created a compilation error with the sign app because
the `function.scss file` was not included in the assets of the module.

[1]: : odoo/odoo@53f605399755f9355554c2582ea0f693c8f947b4

task-3557566

closes odoo/enterprise#52275

X-original-commit: e3a5fe20e1315d5c6329c2de7ef3a63dc9d8b7da
Related: odoo/odoo#145290
Signed-off-by: Pierre Paridans (app) <app@odoo.com>