[FIX] knowledge: prevent filter duplication in embed
Steps to Reproduce
===================
1. Create an article.
2. Insert an embed.
3. Create a default filter from the search menu.
4. Now, create a new article, write the name, and click on 'Edit'.
5. Now, come back to the original article through Breadcrumb.
6. There are multiple filters with the same name.

Technical
==========
The commit that introduces the issue is https://github.com/odoo/enterprise/commit/680171179b06de61ba2bb76aa233598b24427cb6, because before that commit,
an imported state was supposed to come from outside of Knowledge, and therefore
not possess the custom favorite filters.
After that commit, imported states can now sometimes come from inside Knowledge,
but the load override was not adjusted to account for that.

After this commit
==================
Filters will not be visible multiple times in embed.

Task-3562102

closes odoo/enterprise#53211

X-original-commit: 9dad8875f2a30bb1a6bcf7c88bf5c9cc5563736a
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>