[FIX] appointment_account_payment: allow booking contiguous slots
HOW TO REPRODUCE:

With eCommerce and appointment_account_payment installed,
book Mitchell admin on a slot at 9am on a paid appointment.
Once the booking is in cart, book another one at 10pm and with
Mitchell Admin, try to pay the cart. An error message spawns and
warn you that one of the bookings is not available.

ISSUE:

When filtering out unavailable bookings, the availability was
computed on the different time intervals formed by bookings
time ranges. However, the method returning intervals grouped
the contiguous ones. So the user was considered as having a
unique availability on the grouped intervals. Therefore, all
but the first would be considered as unavailable.

FIX:

Use custom means to check all intervals, not merging contiguous
ones. A test is added. At the same time, coverage is maximized
in a simple way for bookings based on users. A condition is also
added for resources when evaluating booking availability to avoid
a crash when a resource is not linked to any appointment anymore.

Task-3570588

closes odoo/enterprise#52962

X-original-commit: bb0e6f155ea52dd41bd380031e91b09def6efc47
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>