[FIX] documents_spreadsheet: fix revision id mismatch
Create a new spreadsheet from a template, then
open "File > See version history".
You'll face the following error:
"There are missing revisions that prevent to restore the whole edition history."

There is some mismatch between revision ids.
Here is what happens:
When we create a spreadsheet from a template:
- we copy the content of `spreadsheet_data` on the `documents.document` record.
  The field now actually contains invalid data for a regular spreadsheet, it
  contains template related stuff.
- when redirecting to the new spreadsheet, the action has an extra param
  `convert_from_template`. Because of this param, we convert the template
  stuff to regular data. This is done by dispatching commands and re-exporting
  the data. Note that dispatching commands updates the revision (even if they
  are not sent to the network)
- finally, we snapshot the spreadsheet to make sure we'll have the converted
  data the next time we open the spreadsheet (and not the data with the template
  stuff)

We end up with a mismatch of revision ids:
the revision id in field `spreadsheet_data` is the revision id from the
original template. The revision id of the snapshot revision is based on the
last dispatched commands (which were not sent to the server).

In this commit, we reset the `spreadsheet_data` field with the converted
data. Otherwise, it will always contains template related stuff which cannot
be open as a regular spreadsheet.
Up until now it didn't translate to any issue because we actually never load
a spreadsheet with the initial `spreadsheet_data` (since it's been snapshotted).
But now, with the new revision history feature, we do need to load the
spreadsheet with the initial data.

Note for existing spreadsheet created from a template which have revision
mismatch: the revision history won't load the entire history (it detects the
mismatch), but it will load the history from the last snapshot.

Task: 3575942
X-original-commit: 0cb2b5169fb1e70da2e7a0efb6f8bc5f67b53ffe
Part-of: odoo/enterprise#50672