[FIX] stock_barcode_mrp: scan lot in series
Usecase to reproduce:
Finished product with 2 components tracked:
- COMP 1 serial A
- COMP 2 serial B
Go on the production order in the barcode app and scan Serial A and
Serial B

Expected Behavior:
Comp 1 1/1 and Comp 2 1/1 and production order to close

Current Behavior:
Comp 1 2/1 with serial A and B

It happens due to the condtion that only recheck if the barcode is
a lot on another product. But it's base on some condition on the
picking type. However for mrp it should not apply

closes odoo/enterprise#50031

X-original-commit: d8952bca704a987d0a191ae8c2cb2db101d59ab0
Signed-off-by: William Henrotin (whe) <whe@odoo.com>