[FIX] account_accountant, web_studio: adapt tests to removal of command "set"
This commit is the enterprise counterpart of https://github.com/odoo/odoo/pull/140644
It simply adapts the tests that should be adapted.

A little comment about the modification in account_accountant:
The step above the one added by this commit was added (https://github.com/odoo/enterprise/pull/47935) in order to
fix a race condition. However, there is still a race condition if the
onchange that add the tax line is not executed before the onchange that
is triggered by the "Validate" button. This is why we add a step that
waits for the tax line to be added. (Runbot-id 24952-24953)

Task-id 3522153

closes odoo/enterprise#50430

X-original-commit: 650f87f293635e2ddbe5865d1e9490162820b161
Related: odoo/odoo#141564
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>