[FIX] sign: sign template edition popover has bigger zindex than modal
In sign when you're editing a template, you can change the role of a sign item
by opening a popover. However, if you click on "search more", it will open a
modal that doesn't overlap the popover due to a bigger zindex of the popover.
This commit fixes this by adjusting the popover's zindex.

On top of this, clicking on a role in the "view all" modal would cause the
modal to close and the clicked role would not be selected. This was being
caused by the closeOnClickAway that would close the popover and modal since any
click outside of the popover would cause it to close it. This commit fixes this
by removing the closeOnClickAway for the sign item popover, as this logic is
already taken care in the sign_template_iframe file.

task-3575711

closes odoo/enterprise#53875

X-original-commit: 5a169e27d2fc6a4c6bac9559c9c52841e88058cc
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>