[FIX] web: clickbot hard-coded date in tests
This commit, fix the issue that some tests fails on other month than
October, this occurs because the current date wasn't patched on the
tests.

Moreover, this commit removes the log "Clicking on: Control Panel menu",
this log is not very important, and can increase significantly the size
of the log.# with '#' will be ignored, and an empty message aborts the
commit.

closes odoo/enterprise#50063

X-original-commit: dbd5672dfb190ba5a46954adb108d302e9c94bb4
Related: odoo/odoo#140774
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>