[FIX] planning: standardize alert design
This commit aims to standardize the design of alerts inside planning by
adding some spacing between alerts and the next element and by
removing custom padding.

Prior to this commit, the alert would be directly stick to the form
sheet, due to the lack of `margin-bottom`. This was causing consistency
issues with other alerts from other modules that would have a bottom
spacing to prevent them to stick to the following element.

There was also a padding property used to add some whitespace within the
alert, while our alert class already include a CSS variables that handles
the padding of the element.

To fix this visual issue, we simply reset the styling applied to these
alert to standard, by removing margin and padding overrides.

task-3580099

closes odoo/enterprise#51118

X-original-commit: 51ac81bee5a92bc301ff9ae195ff657fe72dba83
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>