[IMP] account_reports: Dynamic column width in xlsx
Aims to give a better out-of-the-box layout for xlsx exports
by computing the columns width and height based on the cell content.

Also replace the default font set in this export by Lato instead of Arial.
This way, we are using a font under the SIL Open Font License, which is also
bundled in Odoo meaning that we can use the bundled ttf for our computation purposes.

Task id #3580242

closes odoo/enterprise#50004

Signed-off-by: Quentin De Paoli <qdp@odoo.com>