[IMP] knowledge: update test for get_user_sorted_articles
Introduce a queryCount test for `get_user_sorted_articles` to highlight the
changes in the following commits, which intend to improve performances of this
function.

Alter the `test_article_sort_for_user` business test so that every aspect of the
ordering is actually tested, to ensure that the following changes to
`get_user_sorted_articles` continue to follow the specifications.

task-3554068

X-original-commit: 3dac8bbcc540e8b925aaecf0901b19c957d44f92
Part-of: odoo/enterprise#50482