[REF] pos*: simplify control buttons
The pos app currently uses an overly complicated system
for managing the "control buttons" (the buttons from the left
side of the screen, such as "Refund", "Note" etc. ).

In this PR we refactor these buttons such that we now have a
regular owl component called "ControlButtons" which is
responsible for rendering all the buttons.

Modules can introduce new buttons using the normal extension
mechanisms, such as patching the component.

closes odoo/enterprise#50039

Task: 3580519
Related: odoo/odoo#140723
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>