[FIX] frontdesk: add spacing between image and title
This commit adds a missing space between the image and its title in the
kanban view of Frontdesk.

Prior to this commit, the title of a drink within the kanban view of
frontdesk had no spacing. This was not especially visible since the
default images are using a white background, but this would become an
issue as soon as the image stop having a white or transparent background.

To prevent this unexpected renderer, we simply add a `ms-2` class to
our title, ensuring it will always be spaced correctly.

task-3580511

closes odoo/enterprise#50264

X-original-commit: 43c9fd934d7f97a49cc897c6f6ab01a120954c77
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Chrysanthe Gomrée (chgo) <chgo@odoo.com>