[FIX] timesheet_grid: fix day selection arrow & responsiveness
This commit adapts the timesheet_timer_header to better handle its
responsiveness. On tablet size screen (mg to lg) the view was breaking.

This commit adds a text-nowrap on the stop button and adapt the size of
the inputs next to the timer.

The timer was also moving the content since it's not a monospace font,
a .SCSS file is created to add `tabular-nums` to manage the font size.

task-3580643

closes odoo/enterprise#50055

Related: odoo/odoo#140750
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>