[FIX] timesheet_grid: fix day selection arrow & responsiveness
This commit adapts the timesheet_timer_header to better handle its
responsiveness. On tablet size screen (mg to lg) the view was breaking.

This commit adds a text-nowrap on the stop button and adapt the size of
the inputs next to the timer.

The timer was also moving the content since it's not a monospace font,
a .SCSS file is created to add `tabular-nums` to manage the font size.

task-3580643

closes odoo/enterprise#52054

X-original-commit: 6de4a33d517c9f8ccf431d2b2dbb871430e200dc
Related: odoo/odoo#144830
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>