[FIX] mrp_workorder: shop floor add workorder calculate date_finished
To reproduce:
- Open shop floor
- Open the menu on an MO card and select Add Work Order
- Enter an expected duration and start date and confirm the wizard
- Open the newly created WO in the backend

Current behavior: The end date of the newly created WO is not set when
creating the WO from the shop floor.

Expected behavior: The WO should have an end date calculated from the
start date and expected duration just like when creating it from the
backend.

Task: 3549953 (nr 5)
Part-of: odoo/enterprise#50148